backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
Pillow>=11.0.0
opencv-python>=4.8.0